refactor(packages/contracts): change login and register responses to use userWithPrimaryEmailSchema
